requirements.txt
PyPDF2
ipykernel
google-genai
langchain
langchain-community
python-dotenv
psycopg2-binary